pyproject.toml
[project]
name = "poly-subagents"
version = "0.1.0"
description = "Specialized A2A sub-agents for Polymarket paper trading"
requires-python = ">=3.13"
readme = "README.md"

dependencies = [
    "anthropic>=0.49.0",
    "psycopg2-binary>=2.9.0",
    "weaviate-client>=4.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "schedule>=1.2.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "flask>=3.0.0",
    "waitress>=3.0.0",
    "packaging>=26.1",
    "langfuse>=3.0.0",
    "a2a-sdk[http-server]>=1.0.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "fastapi>=0.136.1",
    "ag-ui-protocol>=0.1.18",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "types-pyyaml>=6.0.0",
    "types-requests>=2.31.0",
    "pytest-asyncio>=0.23.0",
    "types-psycopg2>=2.9.21.20260422",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = [
    "E501",  # line-too-long: line-length=100 è già stretto, evita rumore su print+f-string
]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.13"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
# strict = true richiede grosso type-coverage refactor: abilitiamo progressivamente
ignore_missing_imports = true
disable_error_code = ["import-untyped"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
